Repair the publish script and use the GitHub API to determine whether to tag this `latest` or not (#829)
(cherry picked from commit 11958dadf6b38d14923bba21f5215dcfb894e596)
