+++ packages/build-scripts/constants.ts
@@ -1,0 +1,3 @@
+export const ORG_NAME = 'anza-xyz';
+export const REPO_NAME = 'kit';
+export const SEMVER_REGEX = /^0*(\d+)\.0*(\d+)\.0*(\d+)$/;
